requirements.txt
aiogram
aiohttp
python-dotenv
asyncio
requests
ddinsta
